apply changes, reusability, refactoring
