+++ src/main/java/com/rybka/config/CurrencyAPIConstants.java
@@ -1,4 +1,4 @@
-package com.rybka;
+package com.rybka.config;
 
 public final class CurrencyAPIConstants {
     public static final String PRIME_EXCHANGE_RATE_API_URL = "https://prime.exchangerate-api.com/v5/%s/latest/%s";
