+++ src/main/java/com/rybka/config/PropertyInfo.java
@@ -1,0 +1,6 @@
+package com.rybka.config;
+
+public final class PropertyInfo {
+    public static final String PROPERTY_FILE_NAME = "application.properties";
+    public static final String EXCHANGE_SOURCE = "exchange.source";
+}
